fixes for windows, some things like getopt where missing (added them … …to CMakeLists.txt, found via codeproject: http://www.codeproject.com/Articles/157001/Full-getopt-Port-for-Unicode-and-Multibyte-Microso) and an "Entity" inside PolyPhysicsSceneEntity was ambigous.
Also, some windows headers are defining far & near, so they had to be undefined by the create_lua_library.py script.
Added an inline "round" function in polyglobals for _WINDOWS and changed some usings of "M_PI" to "PI", as they also could not be found by the compiler (even, if math was included).